+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Release date: 2019-02-REPLEACE_ME.
 
 ### Added
 
+- [CLI cancel order command](https://github.com/eidoo/hybrid-exchange-sdk/issues/26)
 - [CLI get allowance command](https://github.com/eidoo/hybrid-exchange-sdk/issues/24)
 - [CLI get trading wallet asset balance command](https://github.com/eidoo/hybrid-exchange-sdk/issues/20)
 - [CLI deposit ETH command](https://github.com/eidoo/hybrid-exchange-sdk/issues/15)